refactor: naming tokenTradingModal -> tokenWarningModal
